Apdate openapi.yaml and completing code coverage

+++ src/main/java/com/gftproject/shoppingcart/services/ShoppingCartServiceImpl.java
@@ -1,5 +1,7 @@
 package com.gftproject.shoppingcart.services;
 
+import com.gftproject.shoppingcart.exceptions.CartNotFoundException;
+import com.gftproject.shoppingcart.exceptions.CartIsAlreadySubmittedException;
 import com.gftproject.shoppingcart.exceptions.NotEnoughStockException;
 import com.gftproject.shoppingcart.exceptions.ProductNotFoundException;
 import com.gftproject.shoppingcart.exceptions.UserNotFoundException;
@@ -56,6 +58,8 @@ public class ShoppingCartServiceImpl implements ShoppingCartService {
 
     @Override
     public Cart createCart(Long userId) throws UserNotFoundException {
+        userService.getUserById(userId);
+
         Cart cart = new Cart();
 
         cart.setUserId(userId);
@@ -67,18 +71,13 @@ public class ShoppingCartServiceImpl implements ShoppingCartService {
     }
 
     @Override
-    public Cart addProductToCartWithQuantity(long userId, long cartId, long productId, int quantity) throws ProductNotFoundException, NotEnoughStockException {
+    public Cart addProductToCartWithQuantity(long cartId, long productId, int quantity) throws ProductNotFoundException, NotEnoughStockException, CartNotFoundException, CartIsAlreadySubmittedException {
         // Check if the cart exists or create a new one if it doesn't
-        Cart cart = cartRepository.findById(cartId).orElseGet(() -> {
-            Cart newCart = new Cart();
-            newCart.setUserId(userId);
-            newCart.setFinalPrice(new BigDecimal(0));
-            newCart.setFinalWeight(new BigDecimal(0));
-            newCart.setStatus(Status.DRAFT);
+        Cart cart = cartRepository.findById(cartId).orElseThrow(() -> new CartNotFoundException("Cart not found" + cartId));
 
-            return cartRepository.save(newCart);
-        });
-
+        if (cart.getStatus().equals(Status.SUBMITTED)){
+            throw new CartIsAlreadySubmittedException(cartId);
+        }
 
         // Check if the product exists
         ProductDTO product = productService.getProductById(productId);
@@ -95,8 +94,8 @@ public class ShoppingCartServiceImpl implements ShoppingCartService {
                 cartProduct = new CartProduct(cart, productId, true, quantity);
             }
 
-            cart.setFinalPrice(cart.getFinalPrice().add(product.getPrice()));
-            cart.setFinalWeight(cart.getFinalWeight().add(product.getWeight()));
+            cart.setFinalPrice(product.getPrice().multiply(BigDecimal.valueOf(quantity)).add(cart.getFinalPrice()));
+            cart.setFinalWeight(product.getWeight().multiply(BigDecimal.valueOf(quantity).add(cart.getFinalWeight())));
             // Save the changes to the cart and cartProduct
             cartRepository.save(cart);
             cartProductRepository.save(cartProduct);
@@ -108,10 +107,15 @@ public class ShoppingCartServiceImpl implements ShoppingCartService {
     }
 
     @Override
-    public Cart submitCart(Long idCart) throws NotEnoughStockException, ProductNotFoundException, UserNotFoundException {
+    public Cart submitCart(Long idCart) throws NotEnoughStockException, ProductNotFoundException, UserNotFoundException, CartIsAlreadySubmittedException {
 
         // Obtain the cart
         Cart cart = cartRepository.findById(idCart).orElseThrow();
+
+        if (cart.getStatus().equals(Status.SUBMITTED)){
+            throw new CartIsAlreadySubmittedException(idCart);
+        }
+
         List<CartProduct> cartProductList = cartProductRepository.findAllByCartId(idCart);
 
         User user = userService.getUserById(cart.getUserId());
@@ -172,10 +176,19 @@ public class ShoppingCartServiceImpl implements ShoppingCartService {
     }
 
     @Override
-    public void deleteProductFromCart(Long cartId, Long productId) {
-        Cart cart = cartRepository.findById(cartId).orElseThrow();
+    public void deleteProductFromCart(Long cartId, Long productId) throws CartNotFoundException, ProductNotFoundException {
+        Cart cart = cartRepository.findById(cartId).orElseThrow(() -> new CartNotFoundException("CART NOT FOUND " + cartId));
+
         CartProduct product = cartProductRepository.findByCartAndProduct(cart, productId);
+        if (product == null){
+            throw new ProductNotFoundException("Product not found in cart: " + productId);
+        }
         cartProductRepository.delete(product);
+    }
+
+    @Override
+    public List<CartProduct> findAllByCartId(Long cartId) {
+        return cartProductRepository.findAllByCartId(cartId);
     }
 
     @Override
